requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
asyncer==0.0.1
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fastapi==0.80.0
filelock==3.8.0
filetype==1.1.0
flatbuffers==22.10.26
frozenlist==1.3.3
gdown==4.5.1
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.2.1
imageio==2.22.4
llvmlite==0.39.1
Markdown==3.4.1
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.8
numba==0.56.4
numpy==1.21.6
onnxruntime==1.12.1
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==4.21.9
pydantic==1.10.2
PyJWT==2.6.0
PyMatting==1.1.8
pyparsing==3.0.9
PySocks==1.7.1
python-multipart==0.0.5
pytz==2022.6
PyWavelets==1.4.1
rembg==2.0.25
requests==2.28.1
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.19.1
sympy==1.11.1
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.0
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
uvicorn==0.18.3
watchdog==2.1.9
yarl==1.8.1
